feat(api-key): add rest api endpoint information (#2079)
* feat(api-key): add rest api endpoint information


* fix: add "Endpoints" to title


---------
